:recycle: Refactor: Move ComResDto.LikeDto to UserHelpInfoResponseDto and delete ComResDto
